standardize behavior of mapping anonymous fields
the behavior has been defined in a way that is compatible with
encoding/json.  this behavior is as follows:

anonymous fields which are structs will have struct fields get
field names as if they were directly in the parent struct.

anonymous fields which are not structs, or which are interfaces
which may or may not point to structs will get field names that
correspond to the name of the type

the exception to the rules above is that you can always override
this behavior by using a JSON struct tag

fixes #101