optimization when boolean query has should constituents only
A common search case is when a user performs a query-string query,
such as for "the lazy dog".  That would be parsed into a boolean query
with a nil Must child, a nil MustNot child, and a non-nil Should child
(a disjunction query for "the", "lazy", "dog").

The optimization in this case is to return just the Should child
directly, skipping any additional Must and MustNot overhead.

On a dev box bleve-query benchmark on a wiki index with a query string
of "text:afternoon text:coffee", the throughput was previously 873qps
and with this change hits 940qps.