fix issue with numeric range queries in query string
previously the query string queries were modified to aid in
compatibility with other search systems.  this change:
f391b991c20f02681bacd197afc6d8aed444e132
has a problem when combined with:
77101ae424bb385095b8094dc9fd1cd116a3de8c
due to the introduction of MatchNoneSearchers being returned
in a case where previously they never would.

the fix for now is to simply return disjunction queries on 0
terms instead.  this ultimately also matches nothing, but avoids
triggering the logic which handles match none searchers in a
special way.