Adapting rollback test to better test the scenario
The rollback test only checks if the new root has an epoch
that is less than or equal to the one reverted to .. which
is pretty weak. Improving this test to write a couple of
batches and retrieve the snapshot between the two and
rollback to that. Now we can better validate that rollback
was successful by ensuring that we only see the data before
the second batch.