FuzzyResult scoring driven by the edit distance
Scoring logic changed to accomodate the edit distance
for the fuzzy query results, so that the exact matches
should rank higher than the other near similar terms.
FuzzyQuery is rewritten as a set of TermQueries that
share a common document frequency of the maximum
document frequency found.
The original boost is multiplied by an edit distance
based boost.
Query coordinate factor is ignored for a fuzzy query.